move to src/lamplib/src to keep things simple
