+++ src/lamplib/src/genny/tasks/compile.py
@@ -2,8 +2,8 @@ from typing import List
 
 import structlog
 import os
-from genny.cmd_runner import run_command
 
+from genny.cmd_runner import run_command
 from genny import toolchain
 
 SLOG = structlog.get_logger(__name__)
